Revamped day 22
The original Day 22 2021 was mostly an implementation of a solution that I found on Reddit, but I was uncomfortable with how similar it was. This is the result of a multistep process of 1) sitting down to understand how the Reddit solution worked, 2) taking a break from the problem and 3) using my recollection of how the Reddit solution worked to reimplement it.