Day 22 2022 code cleanup
Cleaned up names/functions a bit. Cubes are now Cuboids to match the names used in the problem statement (as well as the fact that they were technically not cubes), and run_instructions() (formerly run_instructions_reverse) does its work directly instead of by calling a local function -- that local function had originally existed to keep the outer function's signature cleaner, but the extra parameters turned out to always get the same values as the function was developed further, so it was no longer helpful.