Update utility.d
Only adds a comment documenting an issue with it when the T in Grid2D is a struct and you want to mutate it. Didn't test with classes, but it might work because of the value/reference type semantics.

I'm not sure if it CAN be fixed, since D doesn't seem to allow returning a reference, but there are workarounds, namely manually using the index_at_pt functions as an index into layout.