Explicit handling of FileException in run_day
Instead of the default exception handling, FileException is now explicitly caught so we can print *just the msg* instead of the whole thing.

Usually this is because the file wasn't found, and if not it's probably going to be something like permissions -- these aren't bugs in the program, so printing the backtrace just makes clutter.

As a side effect this no longer halts execution when a FileException occurs when running several days at a time. Arguably this is a bugfix, though that was not the motivation behind this change.